requirements.txt
opencv2-python
tensorflow
numpy